requirements.txt
fastapi
uvicorn
opencv-python
ultralytics
easyocr
pandas
python-Levenshtein
numpy
python-multipart